refactoring: move json functions to emvjson.h/c
